Made BillModel

+++ Data.Entities/Bill.cs
@@ -11,7 +11,7 @@ namespace Data.Entities
         //Primary key
         public int BillId { get; set; }
 
-        public int Amount { get; set; }
+        public float Amount { get; set; }
         public string PaymentStatus { get; set; }
         public DateTime LastUpdate { get; set; }
